requirements.txt
requests
beautifulsoup4
anthropic
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
gspread
